requirements.txt
PyMySQL
fastapi
uvicorn
sqlmodel
cryptography
python-ffmpeg
mutagen
httpx